a little cleaner utilising Feathers-Vuex built in clone, commit, reset methods'
