+++ src/components/UserProfile.vue
@@ -1,27 +1,20 @@
 <template>
-  <Container class="outer-container">
-    <div class="container">
-      <p>Email: {{ email }}</p>
-      <p>Name: {{ name }}</p>
-      <div class="profile-container">
-        Profile Picture:
-        <div class="image-container">
-          <img class="is128x128" :src="profilePicture" />
-        </div>
+  <div class="profile">
+    <p>Email: {{ email }}</p>
+    <p>Name: {{ name }}</p>
+    <div class="profile-container">
+      Profile Picture:
+      <div class="image-container">
+        <img class="is128x128" :src="profilePicture" />
       </div>
-    </div>
-    <UsersList class="users-list" />
-  </Container>
-</template>
+    </div></div
+></template>
 
 <script>
 import Vue from 'vue'
-import Container from '../UIComponents/Container'
-import UsersList from '../components/UsersList'
-import { mapGetters, mapState } from 'vuex'
+import { mapState, mapGetters } from 'vuex'
 export default Vue.extend({
   name: 'UserProfile',
-  components: { Container, UsersList },
   computed: {
     ...mapState('auth', { loading: 'isAuthenticatePending' }),
     ...mapGetters('auth', { user: 'user' }),
@@ -47,15 +40,11 @@ export default Vue.extend({
 })
 </script>
 
-<style scoped>
-.outer-container {
-  align-items: flex-start;
-  justify-content: flex-start;
-}
-.container {
+<style>
+.profile {
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: center;
   align-items: center;
   max-width: 300px;
   height: 400px;
@@ -67,8 +56,5 @@ export default Vue.extend({
 }
 .profile-container {
   flex-direction: row;
-}
-.users-list {
-  width: 100%;
 }
 </style>
